==> notebooks/01_load_and_validate_raw_data.ipynb <==
from pathlib import Path
import requests

def download_one_file_of_raw_data(year : int, month : int) -> Path:
  """"""
  URL = f'https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_2023-01.parquet'
  response = requests.get(URL)

  if response.status_code == 200:
    path = f'../data/raw/rides_{year}-{month:02d}.parquet'
    open(path, "wb").write(response.content)
    return path
  else:
    raise Exception(f'{URL} is not available')

download_one_file_of_raw_data(year=2023, month=1)

import pandas as pd

rides = pd.read_parquet('../data/raw/rides_2023-01.parquet')

rides.head(20)

rides = rides[['tpep_pickup_datetime', 'PULocationID']]

rides = rides.rename(columns={'tpep_pickup_datetime':'pickup_datetime', 'PULocationID':'pickup_location_id'})

rides

rides['pickup_datetime'].describe()

rides = rides[rides.pickup_datetime >= '2023-01-01']
rides = rides[rides.pickup_datetime <  '2023-02-01']
rides.pickup_datetime.describe()

rides.to_parquet('../data/transformed/validated_rides_2023_01.parquet')
