==> notebooks/06_baseline_model.ipynb <==
import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df

from datetime import datetime
from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(
  df,
  cutoff_date=datetime(2023, 6, 1, 0, 0, 0),
  target_column_name='target_rides_next_hour'
)

print(f'{X_train.shape}')
print(f'{y_train.shape}')
print(f'{X_test.shape}')
print(f'{y_test.shape}')

import numpy as np

class BaselineModelPreviousHour:
  """
  Prediction = actual demand observed in the last hour 
  """
  def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
    pass

  def predict(self, X_test: pd.DataFrame) -> np.array:
    """"""
    return X_test[f'rides_previous_1_hour']

model = BaselineModelPreviousHour()
predicts = model.predict(X_test)
predicts

from sklearn.metrics import mean_absolute_error

test_mae = mean_absolute_error(y_test, predicts)
print(f'{test_mae=:.4f}') 

### For week BaseLine Model

class BaselineModelPreviousWeek:
  """
  Predcition : actual demand observed at t -7 days 
  """
  def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
    pass

  def predict(self, X_test: pd.DataFrame) -> np.array:
    """"""
    return X_test[f'rides_previous_{7*24}_hour']

model_week = BaselineModelPreviousWeek()
predicts_week = model_week.predict(X_test)
predicts_week

test_mae_week = mean_absolute_error(y_test, predicts_week)
print(f'{test_mae_week=:.4f}') 

### For Last 4 Weeks Baseline Model

class BaselineModelPreviousLast4Week:
  """
  Predcition : actual demand observed at t - 7 days, t - 14 days, t - 21 days, t - 28 days 
  """
  def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
    pass

  def predict(self, X_test: pd.DataFrame) -> pd.Series:
    """"""
    return 0.25*(
      X_test[f'rides_previous_{7*24}_hour'] + \
      X_test[f'rides_previous_{2*7*24}_hour'] + \
      X_test[f'rides_previous_{3*7*24}_hour'] + \
      X_test[f'rides_previous_{4*7*24}_hour'] 
    )

model_last_4_week = BaselineModelPreviousLast4Week()
predicts_last_4_week = model_last_4_week.predict(X_test)

test_mae_last_4_week = mean_absolute_error(y_test,predicts_last_4_week)
print(f'{test_mae_last_4_week=:.4f}')
